backend/requirements.txt
bcrypt==5.0.0
blinker==1.9.0
click==8.3.2
Flask==3.1.3
flask-cors==6.0.2
Flask-JWT-Extended==4.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.2.3
pandas==2.2.3
PyJWT==2.12.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2025.1
six==1.17.0
tzdata==2025.1
Werkzeug==3.1.8
